pyproject.toml
[project]
description = "Amiga development kit for third party hardware or software extensions."
authors = [
    {name = "Farm-ng Inc."},
    {email = "info@farm-ng.com" }
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.6"

keywords = ["robotics"]

# https://gist.github.com/nazrulworld/3800c84e28dc464b2b30cec8bc1287fc
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development",
    "Framework :: Robot Framework",
]

dependencies = [
    "protobuf==3.20.1",
    "grpcio==1.47.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.1.2",
    "pre-commit==2.20.0",
    "pydoc-markdown==4.6.3",
]

[project.urls]
homepage = "https://github.com/farm-ng/amiga-dev-kit"
documentation = "https://github.com/farm-ng/amiga-dev-kit"
repository = "https://github.com/farm-ng/amiga-dev-kit"

[tool.setuptools]
package-dir = {"" = "lib"}

[tool.setuptools.packages.find]
where = ["lib"]

[tool.setuptools.dynamic]
version = {attr = "farm_ng.__version__"}

[tool.black]
line-length = 120
include = '\.pyi?$'
skip-string-normalization = true
skip-magic-trailing-comma = true

[tool.isort]
profile = "black"
